update for waviness metrics
